Added Enum Translations and Copyrights
Every proto file will now generate an extra "*_model.ts" file which contains implementations of Typescript code instead of just Types for Javascript.
This way, we can use the functionality of the Enum translationTables, which we can use by adding "@Translate" in a Leading comment over a Proto Enum Value.

This functionality will be re-used to allow for Database Doc exportations too.